refractor: valkyrie.plugins.secrets.
