+++ valkyrie/plugins/secrets/__init__.py
@@ -1,13 +1,11 @@
 """
 Valkyrie Secretis Detector Plugin.
 """
-import re
 import hashlib
 from typing import (
-    Set, List, Pattern, Dict, Any
+    List, Dict, Any
 )
 from pathlib import Path
-from dataclasses import dataclass, field
 
 from valkyrie.plugins import BaseSecurityRule
 from valkyrie.core.types import (
@@ -16,19 +14,9 @@ from valkyrie.core.types import (
     FileLocation
 )
 
-
-####
-##      SECRET PATTERN MODEL
-#####
-@dataclass
-class SecretPattern:
-    """Pattern definition for secret detection"""
-
-    name: str
-    pattern: Pattern[str]
-    entropy_threshold: float = 0.0
-    keywords: Set[str] = field(default_factory=set)
-    file_extensions: Set[str] = field(default_factory=set)
+from valkyrie.plugins.secrets.conf import (
+    SECRETS_PATTERNS, SecretPattern
+)
 
 
 ####
@@ -50,38 +38,7 @@ class SecretsDetectionRule(BaseSecurityRule):
         super().__init__(metadata)
         
         # Define secret patterns
-        self.patterns = [
-            SecretPattern(
-                name = "AWS Access Key",
-                pattern = re.compile(r'AKIA[0-9A-Z]{16}', re.IGNORECASE),
-                keywords = {"aws", "amazon", "access", "key"}
-            ),
-            SecretPattern(
-                name = "Generic API Key",
-                pattern = re.compile(r'(?i)(api[_-]?key|apikey|secret[_-]?key|secretkey)\s*[:=]\s*["\']?([a-z0-9]{20,})', re.IGNORECASE),
-                entropy_threshold = 3.5
-            ),
-            SecretPattern(
-                name = "JWT Token",
-                pattern = re.compile(r'eyJ[a-zA-Z0-9_-]*\.eyJ[a-zA-Z0-9_-]*\.[a-zA-Z0-9_-]*'),
-                keywords = {"jwt", "token", "bearer"}
-            ),
-            SecretPattern(
-                name = "GitHub Token",
-                pattern = re.compile(r'gh[pousr]_[A-Za-z0-9_]{36}'),
-                keywords = {"github", "token"}
-            ),
-            SecretPattern(
-                name = "Private Key",
-                pattern = re.compile(r'-----BEGIN [A-Z ]+ PRIVATE KEY-----'),
-                keywords = {"private", "key", "rsa", "ssh"}
-            ),
-            SecretPattern(
-                name = "Database Connection String",
-                pattern = re.compile(r'(mongodb|mysql|postgres|redis)://[^\s]+', re.IGNORECASE),
-                keywords = {"database", "connection", "string"}
-            )
-        ]
+        self.patterns = SECRETS_PATTERNS
     
     def _calculate_entropy(self, text: str) -> float:
         """Calculate Shannon entropy of text"""
